move prepare_pec to pec_utils
